backend/pyproject.toml
[project]
name = "smart-eyes-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "alembic>=1.18.4",
    "cryptography>=46.0.5",
    "fastapi>=0.135.1",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
